- 코드 정리 중 - Spc Configuration 추가 - Spc DeviceManager 개편   -- Attribute 기능 추가   -- 지원 가능한 Device 주소 short -> int 확장

+++ SPC.Core/DataModel/WordReadWriteInfo.cs
@@ -2,7 +2,7 @@
 {
     public class WordReadWriteInfo : DeviceReadWriteInfo
     {
-        public short Size { get; set; }
+        public int Size { get; set; }
         public short[] Value;
     }
 }
